# documentation/api/openapi.yaml
openapi: 3.0.3
info:
  title: Multi-Channel AI Agent API
  description: |
    A scalable multi-channel AI agent service that enables businesses to deploy AI chatbots 
    across WhatsApp, Instagram, and Web Chat platforms.
    
    ## Features
    - Multi-tenant architecture with user_id isolation
    - Support for WhatsApp, Instagram, and Web Chat
    - AI-powered responses using OpenAI
    - Automatic lead extraction and scoring
    - Message delivery tracking
    - Credit-based usage model
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: https://api.example.com
    description: Production server
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /users/{user_id}/phone_numbers:
    post:
      tags:
        - Users
      summary: Add a phone number
      operationId: addPhoneNumber
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
                - platform
                - meta_phone_number_id
                - access_token
              properties:
                id:
                  type: string
                platform:
                  type: string
                  enum: [whatsapp, instagram, webchat]
                meta_phone_number_id:
                  type: string
                access_token:
                  type: string
                display_name:
                  type: string
      responses:
        '201':
          description: Phone number added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  timestamp:
                    type: string
                  correlationId:
                    type: string
    
    get:
      tags:
        - Users
      summary: List phone numbers
      operationId: listPhoneNumbers
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: platform
          in: query
          schema:
            type: string
            enum: [whatsapp, instagram, webchat]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Phone numbers retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                  count:
                    type: integer
  /users/{user_id}/phone_numbers/{phone_number_id}:
    delete:
      tags:
        - Users
      summary: Delete a phone number
      operationId: deletePhoneNumber
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: phone_number_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Phone number deleted successfully
  /users/{user_id}/credits:
    get:
      tags:
        - Users
      summary: Get credit balance
      operationId: getCredits
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Credits retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      user_id:
                        type: string
                      remaining_credits:
                        type: integer
                      last_updated:
                        type: string
  /users/{user_id}/credits/add:
    post:
      tags:
        - Users
      summary: Add credits
      operationId: addCredits
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
              properties:
                amount:
                  type: integer
                  minimum: 1
      responses:
        '200':
          description: Credits added successfully
  /users/{user_id}/agents:
    post:
      tags:
        - Agents
      summary: Create an agent
      operationId: createAgent
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - agent_id
                - phone_number_id
                - prompt_id
                - name
              properties:
                agent_id:
                  type: string
                phone_number_id:
                  type: string
                prompt_id:
                  type: string
                name:
                  type: string
      responses:
        '201':
          description: Agent created successfully
    
    get:
      tags:
        - Agents
      summary: List agents
      operationId: listAgents
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agents retrieved successfully
  /users/{user_id}/agents/{agent_id}:
    get:
      tags:
        - Agents
      summary: Get an agent
      operationId: getAgent
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent retrieved successfully
    
    patch:
      tags:
        - Agents
      summary: Update an agent
      operationId: updateAgent
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                prompt_id:
                  type: string
      responses:
        '200':
          description: Agent updated successfully
    
    delete:
      tags:
        - Agents
      summary: Delete an agent
      operationId: deleteAgent
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent deleted successfully
  /users/{user_id}/messages:
    get:
      tags:
        - Messages
      summary: Get messages
      operationId: getMessages
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: conversation_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Messages retrieved successfully
  /users/{user_id}/messages/stats:
    get:
      tags:
        - Messages
      summary: Get message statistics
      operationId: getMessageStats
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: time_range
          in: query
          schema:
            type: string
            enum: [hour, day, week, month]
            default: day
      responses:
        '200':
          description: Statistics retrieved successfully
  /users/{user_id}/conversations:
    get:
      tags:
        - Messages
      summary: Get conversations
      operationId: getConversations
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: is_active
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Conversations retrieved successfully
  /users/{user_id}/conversations/{conversation_id}:
    get:
      tags:
        - Messages
      summary: Get a conversation
      operationId: getConversation
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: conversation_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Conversation retrieved successfully
  /users/{user_id}/conversations/{conversation_id}/messages:
    get:
      tags:
        - Messages
      summary: Get conversation messages
      operationId: getConversationMessages
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: conversation_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Messages retrieved successfully
  /users/{user_id}/extractions:
    get:
      tags:
        - Extractions
      summary: Get extractions
      operationId: getExtractions
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: latest_only
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Extractions retrieved successfully
  /users/{user_id}/extractions/stats:
    get:
      tags:
        - Extractions
      summary: Get extraction statistics
      operationId: getExtractionStats
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Statistics retrieved successfully
  /users/{user_id}/extractions/export:
    get:
      tags:
        - Extractions
      summary: Export extractions
      operationId: exportExtractions
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        '200':
          description: Extractions exported successfully
  /users/{user_id}/extractions/{extraction_id}:
    get:
      tags:
        - Extractions
      summary: Get an extraction
      operationId: getExtraction
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: extraction_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Extraction retrieved successfully
  /users/{user_id}/conversations/{conversation_id}/extraction:
    get:
      tags:
        - Extractions
      summary: Get conversation extraction
      operationId: getConversationExtraction
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: conversation_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Extraction retrieved successfully
  /users/{user_id}/conversations/{conversation_id}/extract:
    post:
      tags:
        - Extractions
      summary: Trigger extraction
      operationId: triggerExtraction
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: conversation_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Extraction job queued successfully
  /webhook/meta:
    get:
      tags:
        - Webhooks
      summary: Webhook verification
      operationId: verifyWebhook
      parameters:
        - name: hub.mode
          in: query
          required: true
          schema:
            type: string
        - name: hub.verify_token
          in: query
          required: true
          schema:
            type: string
        - name: hub.challenge
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Verification successful
    
    post:
      tags:
        - Webhooks
      summary: Receive webhook
      operationId: receiveWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook received and processed
  /api/cache/invalidate/session:
    post:
      tags:
        - Cache
      summary: Invalidate session cache
      operationId: invalidateSessionCache
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phoneNumberId
                - customerPhone
              properties:
                phoneNumberId:
                  type: string
                customerPhone:
                  type: string
      responses:
        '200':
          description: Cache invalidated successfully
  /api/cache/invalidate/phone-number:
    post:
      tags:
        - Cache
      summary: Invalidate phone number cache
      operationId: invalidatePhoneNumberCache
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phoneNumberId
              properties:
                phoneNumberId:
                  type: string
      responses:
        '200':
          description: Cache invalidated successfully
  /api/cache/invalidate/agent:
    post:
      tags:
        - Cache
      summary: Invalidate agent cache
      operationId: invalidateAgentCache
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - agentId
              properties:
                agentId:
                  type: string
      responses:
        '200':
          description: Cache invalidated successfully
  /api/cache/invalidate/credits:
    post:
      tags:
        - Cache
      summary: Invalidate credits cache
      operationId: invalidateCreditsCache
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
              properties:
                userId:
                  type: string
      responses:
        '200':
          description: Cache invalidated successfully
...
  /api/webchat/{webchat_id}/config:
    get:
      tags:
        - Webchat
      summary: Get configuration page URL
      description: |
        Returns the URL to the visual configuration page where users can customize
        widget colors, preview changes, and copy the embed code.
      operationId: getWebchatConfigPage
      parameters:
        - name: webchat_id
          in: path
          required: true
          schema:
            type: string
          description: Webchat channel identifier
          example: "webchat_user_123_abc"
      responses:
        '200':
          description: Configuration URL retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      webchat_id:
                        type: string
                      config_url:
                        type: string
                        description: URL to open for visual color customization
                        example: "https://api.example.com/widget-config.html?agent_id=webchat_user_123_abc"
                      message:
                        type: string
                        example: "Open this URL to customize widget colors and get embed code"
                  timestamp:
                    type: string
                    format: date-time
        '404':
          description: Webchat channel not found
  /api/webchat/{webchat_id}/embed:
    get:
      tags:
        - Webchat
      summary: Get embed code
      description: |
        Returns the HTML embed code for the webchat widget with default colors.
        Users can modify the color attributes in the returned code.
      operationId: getWebchatEmbedCode
      parameters:
        - name: webchat_id
          in: path
          required: true
          schema:
            type: string
          description: Webchat channel identifier
      responses:
        '200':
          description: Embed code retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      webchat_id:
                        type: string
                      embed_code:
                        type: string
                        description: HTML embed code with color customization support
                  timestamp:
                    type: string
                    format: date-time
        '404':
          description: Webchat channel not found
...